src/requirements.txt
bsedata==0.6.0
transformers
torch
streamlit
rapidfuzz
langchain-community
accelerate
langchain-huggingface
langgraph
gradio